fix: wrong path to include cmsImage type after refactor
