+++ src/actions/groups/committees/Types.ts
@@ -1,5 +1,5 @@
 import type { Committee } from '@prisma/client'
-import type { ExpandedCmsImage } from '@/actions/cms/images/Types'
+import type { ExpandedCmsImage } from '@/cms/images/Types'
 
 export type ExpandedCommittee = Committee & {
     logoImage: ExpandedCmsImage
